% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_funs.R
\name{Methods}
\alias{Methods}
\alias{predict.drbart}
\alias{plot.predict.drbart}
\alias{plot.drbart}
\title{Posterior Predictive Quantities from DR-BART}
\usage{
\method{predict}{drbart}(
  object,
  xpred,
  ygrid,
  type = c("density", "distribution", "quantiles", "mean"),
  quantiles = c(0.025, 0.5, 0.975),
  ...
)

\method{plot}{predict.drbart}(x, CI = FALSE, alpha = 0.05, legend_position, ...)

\method{plot}{drbart}(
  x,
  xpred,
  ygrid,
  type = c("density", "distribution", "quantiles", "mean"),
  quantiles = c(0.025, 0.5, 0.975),
  CI = FALSE,
  alpha = 0.05,
  legend_position,
  ...
)
}
\arguments{
\item{object, x}{If calling the \code{predict} or \code{plot} methods, an
object of class \code{drbart}; else, if calling the \code{predict.plot}
method, an object of class \code{predict.drbart} .}

\item{xpred}{A matrix of points describing which conditional densities /
distributions should be estimated. Rows correspond to different conditional
densities and columns to different covariates.}

\item{ygrid}{A numeric vector of y values at which the conditional density /
distribution should be evaluated.}

\item{type}{Type of predictions to be returned. If \code{'density'}, returns
an estimate of the conditional density functions (pdfs) evaluated at points
in \code{ygrid}. If \code{'distribution'}, returns an estimate of the
conditional distribution functions (cdfs), evaluated at points in
\code{'ygrid'}. If \code{'quantiles'}, returns the quantiles of the
conditional density specified by \code{quantiles}. If \code{mean}, returns
the conditional mean.}

\item{quantiles}{If \code{type = 'quantiles'}, the quantiles of the
conditional densities that should be estimated.}

\item{...}{Ignored.}

\item{CI}{Whether credible intervals should be plotted.}

\item{alpha}{If \code{CI = TRUE}, plots \code{100(1 - alpha)\%} credible
intervals.}

\item{legend_position}{A legend position keyword as accepted by
\code{\link{legend}} or \code{'none'} if the legend should be omitted.}
}
\value{
An object of class \code{predict.drbart}, which is an array
  containing posterior draws of the conditional densities, conditional
  distributions, conditional quantiles, or conditional means, depending on
  the value of `type`. If plotting, the object is invisibly returned.
}
\description{
Compute and plot conditional density functions, distribution functions,
quantiles, and means.
}
\details{
\code{predict.drbart} returns posterior predictive draws of functionals of
the conditional densities and is called by \code{plot.drbart} to generate
appropriate plots. The functional is determined by the value of \code{type};
either entire density or distribution functions, selected quantiles, or means
may be estimated. Because quantiles and means are derived from the relevant
densities themselves, in all cases, prediction is a relatively time intensive
procedure, dependent on the number of posterior samples and on the size of
\code{ygrid}. For this reason, the \code{predict} method returns an object of
class \code{predict.drbart}, which has its own associated method:
\code{plot.predict.drbart}. In this way, plots can be re-generated without
the need to repeatedly call \code{predict.drbart}.

Note: estimated quantities will be inaccurate if \code{ygrid} does not fully
capture the high density regions of the conditional densities.
}
